+++ src/assets/mod.rs
@@ -3,9 +3,9 @@ mod parser;
 
 pub(crate) mod vector;
 
-pub use asset_loader::VelloVectorLoader;
+pub use asset_loader::VelloAssetLoader;
 pub use vector::Vector;
-pub use vector::VelloVector;
+pub use vector::VelloAsset;
 
 pub(crate) use asset_loader::VectorLoaderError;
 pub use parser::*;