+++ src/assets/parser.rs
@@ -1,14 +1,13 @@
-use crate::{assets::vector::Vector, VelloVector};
-use bevy::prelude::*;
+use super::asset_loader::VectorLoaderError;
+use crate::{assets::vector::Vector, VelloAsset};
+use bevy::{prelude::*, utils::Instant};
 use std::sync::Arc;
 use vello::{SceneBuilder, SceneFragment};
 use vello_svg::usvg::{self, TreeParsing};
 
-use super::asset_loader::VectorLoaderError;
-
 /// Deserialize the SVG source XML string from the file
-/// contents buffer represented as raw bytes into a `VelloVector`
-pub fn load_svg_from_bytes(bytes: &[u8]) -> Result<VelloVector, VectorLoaderError> {
+/// contents buffer represented as raw bytes into a `VelloAsset`
+pub fn load_svg_from_bytes(bytes: &[u8]) -> Result<VelloAsset, VectorLoaderError> {
     let svg_str = std::str::from_utf8(bytes)?;
 
     let usvg = usvg::Tree::from_str(svg_str, &usvg::Options::default())?;
@@ -21,39 +20,10 @@ pub fn load_svg_from_bytes(bytes: &[u8]) -> Result<VelloVector, VectorLoaderErro
     let width = usvg.size.width();
     let height = usvg.size.height();
 
-    let vello_vector = VelloVector {
-        data: Vector::Static(Arc::new(scene_frag)),
-        local_transform_bottom_center: compute_local_transform(width, height),
-        local_transform_center: compute_local_transform_center(width, height),
-        width,
-        height,
-    };
-
-    Ok(vello_vector)
-}
-
-/// Deserialize the Lottie source JSON string from the file
-/// contents buffer represented as string into a `VelloVector`
-pub fn load_svg_from_str(svg_str: &str) -> Result<VelloVector, VectorLoaderError> {
-    let bytes = svg_str.as_bytes();
-
-    load_svg_from_bytes(bytes)
-}
-
-/// Deserialize the Lottie source JSON string from the file
-/// contents buffer represented as a str into a `VelloVector`
-pub fn load_lottie_from_bytes(bytes: &[u8]) -> Result<VelloVector, VectorLoaderError> {
-    // Load Lottie JSON bytes with the Velato (bodymovin) parser
-    let composition = vellottie::Composition::from_bytes(bytes)
-        .map_err(|err| VectorLoaderError::Parse(format!("Unable to parse lottie JSON: {err:?}")))?;
-
-    let width = composition.width as f32;
-    let height = composition.height as f32;
-
-    let vello_vector = VelloVector {
-        data: Vector::Animated {
-            original: Arc::new(composition),
-            dirty: None,
+    let vello_vector = VelloAsset {
+        data: Vector::Svg {
+            original: Arc::new(scene_frag),
+            playback_started: Instant::now(),
         },
         local_transform_bottom_center: compute_local_transform(width, height),
         local_transform_center: compute_local_transform_center(width, height),
@@ -65,8 +35,41 @@ pub fn load_lottie_from_bytes(bytes: &[u8]) -> Result<VelloVector, VectorLoaderE
 }
 
 /// Deserialize the Lottie source JSON string from the file
-/// contents buffer represented as a str into a `VelloVector`
-pub fn load_lottie_from_str(json_str: &str) -> Result<VelloVector, VectorLoaderError> {
+/// contents buffer represented as string into a `VelloAsset`
+pub fn load_svg_from_str(svg_str: &str) -> Result<VelloAsset, VectorLoaderError> {
+    let bytes = svg_str.as_bytes();
+
+    load_svg_from_bytes(bytes)
+}
+
+/// Deserialize the Lottie source JSON string from the file
+/// contents buffer represented as a str into a `VelloAsset`
+pub fn load_lottie_from_bytes(bytes: &[u8]) -> Result<VelloAsset, VectorLoaderError> {
+    // Load Lottie JSON bytes with the Velato (bodymovin) parser
+    let composition = vellottie::Composition::from_bytes(bytes)
+        .map_err(|err| VectorLoaderError::Parse(format!("Unable to parse lottie JSON: {err:?}")))?;
+
+    let width = composition.width as f32;
+    let height = composition.height as f32;
+
+    let vello_vector = VelloAsset {
+        data: Vector::Lottie {
+            original: Arc::new(composition),
+            dirty: None,
+            playback_started: Instant::now(),
+        },
+        local_transform_bottom_center: compute_local_transform(width, height),
+        local_transform_center: compute_local_transform_center(width, height),
+        width,
+        height,
+    };
+
+    Ok(vello_vector)
+}
+
+/// Deserialize the Lottie source JSON string from the file
+/// contents buffer represented as a str into a `VelloAsset`
+pub fn load_lottie_from_str(json_str: &str) -> Result<VelloAsset, VectorLoaderError> {
     let bytes = json_str.as_bytes();
 
     load_lottie_from_bytes(bytes)