+++ src/plugin.rs
@@ -2,7 +2,7 @@ use crate::{
     font::VelloFont,
     renderer::VelloRenderPlugin,
     rendertarget::{self, SSRT_SHADER_HANDLE},
-    VelloFontLoader, VelloVector, VelloVectorLoader,
+    VelloAsset, VelloAssetLoader, VelloFontLoader,
 };
 use bevy::{asset::load_internal_asset, prelude::*, sprite::Material2dPlugin};
 
@@ -23,8 +23,8 @@ impl Plugin for VelloPlugin {
                 #[cfg(feature = "debug")]
                 crate::debug::DebugVisualizationsPlugin,
             ))
-            .init_asset::<VelloVector>()
-            .init_asset_loader::<VelloVectorLoader>()
+            .init_asset::<VelloAsset>()
+            .init_asset_loader::<VelloAssetLoader>()
             .init_asset::<VelloFont>()
             .init_asset_loader::<VelloFontLoader>()
             .add_systems(Startup, rendertarget::setup_ss_rendertarget)
@@ -35,5 +35,8 @@ impl Plugin for VelloPlugin {
                     rendertarget::clear_when_empty,
                 ),
             );
+
+        #[cfg(feature = "state-machines")]
+        app.add_plugins(crate::state_machine::StateMachinePlugin);
     }
 }