+++ CHANGELOG.md
@@ -6,8 +6,15 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 ## Unreleased
 
+### added
+
+- State machines are now available behind the `state-machines` cargo feature
+- `PlaybackSettings` can now be bundled with `VelloAssetBundle` to augment playback
+
 ### changed
 
+- `RenderMode` changed to `CoordinateSpace`
+- `VelloVector`, anywhere mentioned, has changed to `VelloAsset`
 - `DebugVisualizations` are now feature-gated behind the `debug` cargo feature
 - Color swapping now swaps by layer name only and applies to more cases (animated, gradients, etc.)